Move to ES6
This is a huge commit which changes the internals drastically. I moved
from plain JS to ES6, which will be transpiled by babel to JS.

The lib contains now a gulp file to run some task like gulp babel or
gulp watch.

The template folder has moved to root path. The source root is set
dynamically to fit the generator and its subgenerators.